+++ src/common.cpp
@@ -55,6 +55,8 @@
 #include "achievement.h"
 #include "game_cfg.h"
 
+#include "ff7/widescreen.h"
+
 #include "ff8/vram.h"
 #include "ff8/vibration.h"
 
@@ -835,6 +837,8 @@ int common_create_window(HINSTANCE hInstance, struct game_obj* game_object)
 					vram_init();
 					vibration_init();
 				}
+
+				if(aspect_ratio == AR_WIDESCREEN) widescreen.init();
 
 				// enable verbose logging for FFMpeg
 				av_log_set_level(AV_LOG_VERBOSE);