+++ src/ff7/widescreen.h
@@ -23,6 +23,9 @@
 
 #pragma once
 
+#include "common.h"
+#include "globals.h"
+
 int wide_viewport_x = -106;
 int wide_viewport_y = 0;
 int wide_viewport_width = 854;
@@ -34,3 +37,48 @@ int wide_game_width = 960;
 int wide_game_height = 480;
 
 void ff7_widescreen_hook_init();
+void ff7_field_draw_gray_quads_sub_644E90();
+
+enum WIDESCREEN_MODE
+{
+    WM_DISABLED,
+    WM_EXTEND_ONLY,
+    WM_ZOOM,
+};
+
+class Widescreen
+{
+public:
+    void init();
+    void initParamsFromConfig();
+    void exportConfig();
+    void reloadConfig();
+
+    const field_camera_range& getCameraRange();
+    WIDESCREEN_MODE getMode();
+
+private:
+    void loadConfig();
+
+private:
+    // Config
+    toml::parse_result config;
+
+    field_camera_range camera_range;
+    WIDESCREEN_MODE widescreen_mode = WM_DISABLED;
+};
+
+inline const field_camera_range& Widescreen::getCameraRange()
+{
+    return camera_range;
+}
+
+inline WIDESCREEN_MODE Widescreen::getMode()
+{
+    struct game_mode* mode = getmode_cached();
+    if (mode->driver_mode != MODE_FIELD) return WM_DISABLED;
+
+    return widescreen_mode;
+}
+
+extern Widescreen widescreen;